+++ packages/ko/src/types.ts
@@ -1,6 +1,7 @@
 import { Pattern } from 'copy-webpack-plugin';
 import { Plugin } from 'postcss';
 import { IKeys, IOpts } from 'ko-lints';
+import { IOpts as AutoPolyfillsWebpackPluginOptions } from 'auto-polyfills-webpack-plugin';
 
 export type IOptions = {
   //common configs
@@ -19,6 +20,9 @@ export type IOptions = {
   extraPostCSSPlugins?: Plugin[];
   lessOptions?: any;
   antdV4LessOptions?: any;
+  // integrated plugins options
+  dynamicResolve?: <T extends any>(request: T) => T;
+  autoPolyfills: boolean | AutoPolyfillsWebpackPluginOptions;
   // dev, or serve configs
   serve: {
     proxy?: Record<string, any>;